requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.2.1
feedparser==6.0.11
filelock==3.17.0
gitdb==4.0.12
GitPython==3.1.44
googlesearch-python==1.3.0
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.2
pandas==2.2.3
phidata==2.7.9
pillow==11.1.0
primp==0.10.1
pycountry==24.6.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
tldextract==5.1.3
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0